Update typings for compat with @types/webpack
Break webpack.d.ts into individual declarations beneath	the `types`
directory, and update usage for compatibility with @types/webpack.
